rename base to spaceStation and baseStorage to spaceStationStorage
